+++ packages/ast-manipulation/js/function.ts
@@ -1,4 +1,4 @@
-import { AstKinds, AstTypes } from "@svelte-add/ast-tooling";
+import type { AstKinds, AstTypes } from "@svelte-add/ast-tooling";
 
 export function call(name: string, args: string[]) {
     const callExpression: AstTypes.CallExpression = {