+++ packages/testing-library/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @svelte-add/testing-library
 
+## 2.0.13
+
+### Patch Changes
+
+-   chore: enforce stricter eslint rules ([#378](https://github.com/svelte-add/svelte-add/pull/378))
+
+-   Updated dependencies [[`a0447e833cab3a81908fc6fe4e774b597b65f615`](https://github.com/svelte-add/svelte-add/commit/a0447e833cab3a81908fc6fe4e774b597b65f615), [`128be6cb88d382744c61d4eb2efe595b46ff2dc0`](https://github.com/svelte-add/svelte-add/commit/128be6cb88d382744c61d4eb2efe595b46ff2dc0), [`d5953e96946a2c862957d1f88c52c6a1f8149049`](https://github.com/svelte-add/svelte-add/commit/d5953e96946a2c862957d1f88c52c6a1f8149049), [`5d70e3ebdec06679f512eca65e1739ccfdbc6f6d`](https://github.com/svelte-add/svelte-add/commit/5d70e3ebdec06679f512eca65e1739ccfdbc6f6d), [`f3666c4405799cf936d49b9327f9f8728fd1a35e`](https://github.com/svelte-add/svelte-add/commit/f3666c4405799cf936d49b9327f9f8728fd1a35e)]:
+    -   @svelte-add/core@2.5.1
+    -   svelte-add@2.0.12
+
 ## 2.0.12
 
 ### Patch Changes